% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_NTLKwIEx.R
\name{E_NTLKwIEx}
\alias{E_NTLKwIEx}
\title{Estimate parameters with constraints}
\usage{
E_NTLKwIEx(data)
}
\arguments{
\item{data}{Numeric vector of data values.}
}
\value{
Numeric vector of estimated parameters.
}
\description{
This function estimates the parameters of the NTLKwIEx distribution while adhering to parameter constraints. It employs the maximum likelihood estimation method and returns estimated values for each parameter based on a given dataset and the specified constraints.
}
