% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_NTLKwIEx.R
\name{P_NTLKwIEx}
\alias{P_NTLKwIEx}
\title{Probability Density Function (PDF) of the NTLKwIEx distribution}
\usage{
P_NTLKwIEx(x, teta, alpha, a, b, m)
}
\arguments{
\item{x}{Value to evaluate the PDF at}

\item{teta}{Parameter teta of the distribution}

\item{alpha}{Parameter alpha of the distribution}

\item{a}{Parameter a of the distribution}

\item{b}{Parameter b of the distribution}

\item{m}{Parameter m of the distribution}
}
\value{
Value of the PDF for the NTLKwIEx distribution evaluated at x
}
\description{
The Probability Density Function (PDF) of the NTLKwIEx distribution is defined as:
}
\details{
\deqn{
f(x, \theta, \alpha, a, b, m) =2abm\frac{\theta}{x^{2}}\left( -\frac{\theta}{x}log\left( \alpha \right)+ exp\left( \dfrac{\theta}{x}\right)\right)exp\left\lbrace -\frac{\theta}{x}\left(1+a\alpha^{exp\left(-\frac{\theta}{x}\right)}\right)\right\rbrace \left\lbrace 1-exp\left( -a\frac{\theta}{x}\alpha^{exp\left(-\frac{\theta}{x}\right)} \right)\right\rbrace^{2b-1}\left\lbrace  1-\left\lbrace   1-exp\left( -\frac{a\theta}{x} \alpha^{exp\left(-\frac{\theta}{x}\right) }\right)\right\rbrace ^{2b} \right\rbrace^{m-1}
}
}
