% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CNTLKwIEx.R
\name{Plot_CNTLKwIEx}
\alias{Plot_CNTLKwIEx}
\title{Graphical representation of the Cumulative Distribution Function (CDF) of the NTLKwIEx distribution}
\usage{
Plot_CNTLKwIEx(teta, alpha, a, b, m, min_x, max_x)
}
\arguments{
\item{teta}{Parameter teta of the distribution}

\item{alpha}{Parameter alpha of the distribution}

\item{a}{Parameter a of the distribution}

\item{b}{Parameter b of the distribution}

\item{m}{Parameter m of the distribution}

\item{min_x}{Minimum value of x for the plot}

\item{max_x}{Maximum value of x for the plot}
}
\value{
A plot of the CDF of the NTLKwIEx distribution
}
\description{
This function generates a plot of the Cumulative Distribution Function (CDF) of the NTLKwIEx distribution
over a specified range of x values.
}
