% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_PNTLKwIEx.R
\name{Plot_PNTLKwIEx}
\alias{Plot_PNTLKwIEx}
\title{Graphical representation of the probability density function (PDF) of the NTLKwIEx distribution}
\usage{
Plot_PNTLKwIEx(teta, alpha, a, b, m, min_x, max_x)
}
\arguments{
\item{teta}{Parameter teta of the distribution}

\item{alpha}{Parameter alpha of the distribution}

\item{a}{Parameter a of the distribution}

\item{b}{Parameter b of the distribution}

\item{m}{Parameter m of the distribution}

\item{min_x}{Minimum value of x for the graph}

\item{max_x}{Maximum value of x for the graph}
}
\value{
A graph of the PDF of the NTLKwIEx distribution
}
\description{
This function generates a graph of the probability density function (PDF) of the NTLKwIEx distribution
over a specified range of x values.
}
