% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_NTLKwIEx.R
\name{Q_NTLKwIEx}
\alias{Q_NTLKwIEx}
\title{Quantile Value of the NTLKwIEx distribution}
\usage{
Q_NTLKwIEx(p, teta, alpha, a, b, m)
}
\arguments{
\item{p}{Probability for which the quantile value is to be calculated (0 <= p <= 1)}

\item{teta}{Parameter teta of the distribution}

\item{alpha}{Parameter alpha of the distribution}

\item{a}{Parameter a of the distribution}

\item{b}{Parameter b of the distribution}

\item{m}{Parameter m of the distribution}
}
\value{
The quantile value corresponding to the probability p for the NTLKwIEx distribution
}
\description{
This function calculates the quantile value of the NTLKwIEx distribution
for a given probability p.
}
