% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_TLKwIEx.R
\name{R_NTLKwIEx}
\alias{R_NTLKwIEx}
\title{Random Sampling from the NTLKwIEx distribution}
\usage{
R_NTLKwIEx(n, teta, alpha, a, b, m)
}
\arguments{
\item{n}{Number of random samples to generate}

\item{teta}{Parameter teta of the distribution}

\item{alpha}{Parameter alpha of the distribution}

\item{a}{Parameter a of the distribution}

\item{b}{Parameter b of the distribution}

\item{m}{Parameter m of the distribution}
}
\value{
A vector of n random samples from the NTLKwIEx distribution
}
\description{
This function generates random samples from the NTLKwIEx distribution
based on the given parameters.
}
