% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_NTLKwIEx.R
\name{Sim_NTLKwIEx}
\alias{Sim_NTLKwIEx}
\title{Estimate parameters with constraints}
\usage{
Sim_NTLKwIEx(data)
}
\arguments{
\item{data}{Numeric vector of data values.}
}
\value{
Numeric vector of estimated parameters.
}
\description{
This function generates a histogram that depicts the distribution of the provided input data. Additionally, it estimates the parameters of a distribution that would correspond to the given data. By overlaying the estimated density function onto the histogram, Sim_NTLKwIEx enables an immediate comparison between the empirical distribution and the estimated one. Sim_NTLKwIEx proves to be a valuable tool for initial data exploration, streamlining trend identification, and understanding key features. Its usage comes recommended for tasks that require a swift exploratory analysis of data distributions.
}
\examples{
Sim_NTLKwIEx(c(38.181,	38.542,	38.928,	39.334,35.8))

}
